support tuple/vector of arrays for mosaicview -- a slow version
It's convenient to visualize a list/tuple of images of different
sizes.

The pipeline isn't very efficient, for example, 2d images are
concatenated to 3d images, and then piped into previous 3d mosaicview
methods. In future we can directly implemente a faster version without
changing dimensions, but for now, it's still okay to use this slow
version to ease some of visualization work.

Said that, it's still possible to do a manual cat if performance
is the issue.