center-shift padded images
When images are padded, they stay at the top-left cornor, but for
visualization purpose, it's often convenient to shift images to
the center place

Note that the centered version is slower than the non-centered version,
but still, I think it's worth making it as default behavior.
We don't really care about 0.1ms, 1ms, 2ms or 10ms difference when
displaying images.