Improve inferrability and distinguish `mosaic` from `mosaicview` (#28)
* Improve inferrability

This uses lispy tuple programming to avoid heterogenous broadcasts.
It also breaks the computation into two pieces: promoting the eltypes
of the arrays, and then adding the fill.

This does require changing the result of one test, but the change
appears to increase consistency.

* Reduce specialization for heterogenous inputs

When called with different input array types, it doesn't seem
necessary to specialize since the inference quality will be poor.

* Distinguish `mosaic` from `mosaicview`

The old `mosaicview(A, B)` didn't actually construct a view. This changes
the name to `mosaic` for greater clarity about the relationship between
the inputs and output.